+++ Changelog.md
@@ -1,5 +1,7 @@
 # Changes
 
+* 2016-11-28 \defineSection command support
+* 2016-11-28 Test subsystem
 * 2016-11-27 To make things easier, dpANS sources are now a part of the repo
 * 2016-11-26 Added framework to define commands, basic functions to generate document from stream of commands
 * 2016-11-23 Parser can process the whole file now